third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-language-override-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property font-language-override value 'normal'
  assert_true: font-language-override doesn't seem to be supported in the computed style expected true got false
[FAIL] Property font-language-override value '"KSW"'
  assert_true: font-language-override doesn't seem to be supported in the computed style expected true got false
[FAIL] Property font-language-override value '"ENG "'
  assert_true: font-language-override doesn't seem to be supported in the computed style expected true got false
[FAIL] Property font-language-override value '"en  "'
  assert_true: font-language-override doesn't seem to be supported in the computed style expected true got false
[FAIL] Property font-language-override value '" en "'
  assert_true: font-language-override doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
